Handle overloaded records with popup && now Assessment can receive "delete record" message from Recorder #81

+++ applications/apps/IngeScape-Assessments/model/recordassessmentm.h
@@ -30,7 +30,7 @@ class RecordAssessmentM : public RecordM
     I2_QML_PROPERTY(int, startTimeInTimeline)
 
     // End time in milliseconds of our record in timeline
-    I2_QML_PROPERTY(qint64, endTimeInTimeline)
+    I2_QML_PROPERTY(int, endTimeInTimeline)
 
 public:
     explicit RecordAssessmentM(QString uid,
